fixed some warnings/errors

+++ test/test_lexer.c
@@ -1,3 +1,6 @@
+// NOLINTBEGIN(misc-include-cleaner)
+// we checked to make sure only criterian related warnings were left
+
 #include <criterion/criterion.h>
 #include <criterion/new/assert.h>
 #include <stdio.h>
@@ -38,7 +41,7 @@ Test(lexer, if_whitespace) {
   initLexer(&lexer, source);
   Token token = getNextToken(&lexer);
   cr_assert_eq(token.type, TOKEN_IF);
-  cr_assert(memcmp(token.lexeme, "if", token.length) == 0,
+  cr_assert(memcmp(token.lexeme, "if", (size_t)token.length) == 0,
             "Expected token lexeme to be \"if\"");
   cr_assert_eq(token.length, 2);
   free(source);
@@ -141,7 +144,7 @@ Test(lexer, identifier_with_underscore_and_digits_file) {
   Token token = getNextToken(&lexer);
   cr_assert_eq(token.type, TOKEN_IDENTIFIER);
   cr_assert_eq(token.length, 9);
-  cr_assert(strncmp(token.lexeme, "_myVar123", token.length) == 0);
+  cr_assert(strncmp(token.lexeme, "_myVar123", (size_t)token.length) == 0);
 
   free(source);
 }
@@ -149,7 +152,8 @@ Test(lexer, identifier_with_underscore_and_digits_file) {
 // // Test that verifies skipping a single-line comment via file input
 // Test(lexer, comment_skipping_file) {
 //   char* source = read_file(
-//       CMAKE_SOURCE_DIR "/test/test_inputs/lexer_inputs/comment_skipping.txt");
+//       CMAKE_SOURCE_DIR
+//       "/test/test_inputs/lexer_inputs/comment_skipping.txt");
 //   Lexer lexer;
 //   initLexer(&lexer, source);
 
@@ -170,7 +174,7 @@ Test(lexer, invalid_exclamation_file) {
 
   Token token = getNextToken(&lexer);
   cr_assert_eq(token.type, TOKEN_UNKNOWN);
-  cr_assert(strncmp(token.lexeme, "Unexpected '!'", token.length) == 0);
+  cr_assert(strncmp(token.lexeme, "Unexpected '!'", (size_t)token.length) == 0);
 
   free(source);
 }
@@ -208,7 +212,8 @@ Test(lexer, multiline_tokens_and_line_count_file) {
 // // Test that verifies mixed comments + tokens via file input
 // Test(lexer, multiple_comments_and_tokens_file) {
 //   char* source = read_file(
-//       CMAKE_SOURCE_DIR "/test/test_inputs/lexer_inputs/multiple_com_tok.txt");
+//       CMAKE_SOURCE_DIR
+//       "/test/test_inputs/lexer_inputs/multiple_com_tok.txt");
 //   Lexer lexer;
 //   initLexer(&lexer, source);
 
@@ -239,3 +244,4 @@ Test(lexer, keyword_as_prefix_identifier_file) {
 }
 
 // NOLINTEND(*-magic-numbers)
+// NOLINTEND(misc-include-cleaner)
